+++ app/routes.ts
@@ -4,4 +4,5 @@ export default [
   index("routes/home.tsx"),
   route("comparison", "routes/comparison.tsx"),
   route("are-we-recharts-yet", "routes/are-we-recharts-yet.tsx"),
+  route("styling", "routes/styling.tsx"),
 ] satisfies RouteConfig;